Cleanup
Tweaked spacing to be my preference and consistent.

Changed the Python ifdef stuff to always by Python 3+.
Since all the Python stuff is getting stripped out anyway this barely matters and is just done to make it all more readable as I do that.

Added a gitignore for jedit temporary files.